net: wireless: bcmdhd: do not allow race condition
Module "init" turns off wifi chip after registering
the network interface. The chip is turned on at "open".
As "open" is allowed as soon as registration is
complete, the chip turning on and off becomaes a race
condition. We need to enforce a strict ordering between
these two operations.

Bug 1192094

Change-Id: I17e7da0252e1dd5c335722f98e9b54b0585b9690
Signed-off-by: Mursalin Akon <makon@nvidia.com>
Reviewed-on: http://git-master/r/172442
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>